Fix play button showing when not host
My theory on the bug is that the play button is shown when they've created a room before and joined a new room after. Since the Room object is not recreated, the LocalPlayerIsHost remains true from the time they created the room. This is the behaviour I've observed on Scrapie, Spammie and Wirtual's streams.

This fix assumes that JoinRoom is only called for non-host players. Let me know if that isn't the case!